python/requirements.txt
pybind11>=2.10.0
numpy>=1.19.0
Pillow>=8.0.0
